@@ -1,16 +1,19 @@
 {
-  "name": "maze-generator",
+  "name": "maze-solvers",
   "version": "0.1.0",
   "private": true,
-  "homepage": "http://primaryobjects.github.io/maze-generator",
+  "homepage": "http://fcolt.github.io/maze-solvers",
   "dependencies": {
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
+    "gh-pages": "^6.1.1",
     "react-scripts": "0.9.5"
   },
   "scripts": {
+    "predeploy": "npm run build",
+    "deploy": "gh-pages -d build",
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
